abstract filters
