+++ src/library/Filter/Category.tsx
@@ -1,0 +1,19 @@
+// Copyright 2022 @paritytech/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import React from 'react';
+
+export const Category = ({
+  title,
+  children,
+}: {
+  title: string;
+  children: React.ReactNode;
+}) => {
+  return (
+    <div className="category">
+      <div className="head">{title}</div>
+      <div className="items">{children}</div>
+    </div>
+  );
+};
